+++ app/page.tsx
@@ -1,21 +1,5 @@
 "use client";
 import NavBar from "@/components/ui/nav-bar";
-import "@blocknote/core/fonts/inter.css";
-import { BlockNoteView } from "@blocknote/mantine";
-import "@blocknote/mantine/style.css";
-import {
-  BlockNoteSchema,
-  defaultBlockSpecs,
-  filterSuggestionItems,
-} from "@blocknote/core";
-import {
-  useCreateBlockNote,
-  SuggestionMenuController,
-  getDefaultReactSlashMenuItems,
-  DefaultReactSuggestionItem,
-} from "@blocknote/react";
-import { MermaidBlock } from "@/components/blocks/MermaidBlock";
-import { FaProjectDiagram } from "react-icons/fa";
 import { useTheme } from "@/components/providers/theme-provider";
 import { 
   ResizableHandle, 
@@ -25,68 +9,18 @@ import {
 import { TerminalPanel } from "@/components/ui/TerminalPanel";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { useState, useRef, useEffect } from "react";
-import { useAppStore } from "@/stores/app";
+import dynamic from 'next/dynamic';
+
+// Dynamically import the Editor component with SSR disabled
+const Editor = dynamic(
+  () => import('@/components/Editor'),
+  { ssr: false }
+);
 
 export default function Home() {
-  const schema = BlockNoteSchema.create({
-    blockSpecs: {
-      ...defaultBlockSpecs,
-      mermaid: MermaidBlock,
-    },
-  });
-
   const { theme } = useTheme();
   const [isExpanded, setIsExpanded] = useState(false);
   const panelRef = useRef<any>(null);
-  const { updateUserInput } = useAppStore();
-  
-  // Create editor with custom schema
-  const editor = useCreateBlockNote({
-    schema,
-    initialContent: [
-      {
-        type: "heading",
-        props: {
-          level: 1,
-        },
-        content: "Threat Model",
-      },
-    ],
-  });
-
-  // Create custom slash menu items
-  const getCustomSlashMenuItems = (editor: any): DefaultReactSuggestionItem[] => {
-    // Get the default slash menu items
-    const defaultItems = getDefaultReactSlashMenuItems(editor);
-    
-    // Add custom mermaid item
-    const mermaidItem = {
-      title: "Mermaid Diagram",
-      onItemClick: () => {
-        const currentBlock = editor.getTextCursorPosition().block;
-        // Insert a new mermaid block after the current block
-        editor.insertBlocks(
-          [
-            {
-              type: "mermaid",
-              props: {
-                code: "graph TD;\nA-->B;\nA-->C;\nB-->D;\nC-->D;",
-                textAlignment: "left",
-              },
-            },
-          ],
-          currentBlock,
-          "after"
-        );
-      },
-      aliases: ["diagram", "mermaid", "flowchart", "chart"],
-      group: "Diagrams",
-      icon: <FaProjectDiagram size={18} />,
-      subtext: "Insert a Mermaid diagram",
-    };
-    
-    return [...defaultItems, mermaidItem];
-  };
 
   // Use the imperative API to toggle panel expansion
   const togglePanel = () => {
@@ -116,20 +50,6 @@ export default function Home() {
     }
   }, []);
 
-  // Update Zustand store when editor content changes
-  const handleEditorChange = async () => {
-    // Convert editor blocks to markdown
-    const markdown = await editor.blocksToMarkdownLossy(editor.document);
-    console.log(markdown);
-    // Update description in the store
-    updateUserInput(markdown, "description");
-  };
-
-  // Initial content update
-  useEffect(() => {
-    handleEditorChange();
-  }, []);
-
   return (
     <div className="flex flex-col h-full">
       <NavBar />
@@ -137,19 +57,7 @@ export default function Home() {
         <ResizablePanel defaultSize={75} minSize={30}>
           <div className="p-2 h-full">
             <ScrollArea className="h-[calc(100vh-100px)]">
-              <BlockNoteView 
-                editor={editor} 
-                slashMenu={false}
-                theme={theme === "dark" ? "dark" : "light"}
-                onChange={handleEditorChange}
-              >
-                <SuggestionMenuController
-                  triggerCharacter="/"
-                  getItems={async (query) => 
-                    filterSuggestionItems(getCustomSlashMenuItems(editor), query)
-                  }
-                />
-              </BlockNoteView>
+              <Editor />
             </ScrollArea>
           </div>
         </ResizablePanel>
